requirements.txt
fabric
pymongo
django==1.3
south
bottle
requests
apachelog
pyyaml
#pystache
-e git://github.com/defunkt/pystache.git#egg=pystache
#django-piston
mimeparse>=0.1.3
python-dateutil>=2.1
django-tastypie
PIL
django-bootstrap-form
stripe